Overall project restructuration
I made big changes to the project's overall structure today, in an effort to make the project more maintanable for us all and more understandable for future collaborators (sorry Onur, should've made that before you hopped in :p).

## __`Phase 1` : internal restructuration__
(doesn't affect the rest of the project too much)
- Added comments that reference documentation and explain the code better
- Renamed elements to be more explicit (hence the small changes in `tests/test_phase1.py`)
- Added customized exceptions, and improved modularity overall

## __`Phase 3` : entire restructuration__
- Moved the app startup to `src/main.py`. Now, to launch the app, `cd` to the project's folder then type `python3 src/main.py`.
- Removed `phase3.py` entirely, to introduce 4 new files in a new folder : `src/UI_Functional` ! Note that **none of the actual classes have changed in any manner**, it's just a file and import reordering. Here are the 4 files' purposes :
	- `utilities.py` : contain "standalone" UI components (like a customized `QComboBox`, the `HelpWindow` in our app, etc)
	- `main_window.py` : contains the "skeleton" of our app (e.g the `MainWindow` class) that's in charge of creating the first window you see when opening the app, as well as managing those which follow.
	- `stacked_widget_elems.py` : contains the classes for each "step" the user's gonna go through when using the app (the classes' names are explicit enough)
	- `comparison_results` : contains the results of the computations made in `phase1.py`, absolutely fundamental to display the final results using the `Step4_DisplayResults` class in `stacked_widget_elems.py`.
