+++ src/TextAnalysis_AkaPhase1.py
@@ -1,7 +1,6 @@
-#TODO : support different file types (like .pdf, .md, etc.)
-
 from string import punctuation # helps to check if a string is made of punctuation only here
 from math import sqrt
+import nltk
 from numpy import array, append, flexible, ndarray, mean, median
 from nltk.tokenize import simple, wordpunct_tokenize
 from nltk.corpus import stopwords, wordnet
@@ -12,8 +11,6 @@ from collections import Counter
 from nltk.util import ngrams
 from os import path
 import chardet
-import nltk
-#TODO : fix phase1 to send more concrete and 'actionable' warnings when needed (like when a file is empty)
 
 @dataclass
 class Tokenizer:
@@ -21,28 +18,21 @@ class Tokenizer:
     _tokens_by_sentence: ndarray = field(init=False, repr=False)
     _tokens_by_wordpunct: ndarray = field(init=False, repr=False)
     _tokens_by_word: ndarray = field(init=False, repr=False, default_factory=lambda: array([]))
-    _tokens_by_syntagm: ndarray = field(init=False, repr=False, default_factory=lambda: array([]))
+    _tokens_by_syntagm: ndarray = field(init=False, repr=False, default_factory=lambda: array([])) #syntagm = sequence of contiguous words that are not stopwords, punctuation or empty strings
     _part_of_speeches: ndarray = field(init=False, repr=False, default_factory=lambda: array([]))
 
     def __post_init__(self):
-        # Tokenisation à l'initialisation
         sentences = self._raw_data.replace('\n', 'þ').replace('.', 'þ').split('þ')
         self._tokens_by_sentence = array([sent.strip() for sent in sentences if sent.strip()])
-        self._tokens_by_wordpunct = array(wordpunct_tokenize(self._raw_data)) #first pass, will be filtered in self.clean_data() -> self._filter_and_group_tokens(...)
+        tokens_by_wordpunct_first_pass = array(wordpunct_tokenize(self._raw_data)) #first pass using the 'automatic' method from the nltk lib, result will be filtered in the following lines
+        pos_tags_first_pass = self._lemmatize_and_tag_tokens(tokens_by_wordpunct_first_pass) #-> list[(word, pos_tag)]
+        self._tokens_by_syntagm, self._tokens_by_wordpunct = self._filter_and_group_tokens(pos_tags_first_pass)
 
-        # Nettoyage des données
-        self.clean_data()
 
-    def clean_data(self):
-        """Méthode principale pour nettoyer les données."""
-        pos_tags = self._lemmatize_and_tag_tokens()
-        self._filter_and_group_tokens(pos_tags)
-
-    def _lemmatize_and_tag_tokens(self):
-        """Lemmatization et étiquetage des POS."""
+    def _lemmatize_and_tag_tokens(self, tbwp_first_pass) -> list[tuple[str, str]]:
         lemmatizer = WordNetLemmatizer()
         pos_tags = []
-        for token, tag in pos_tag(self._tokens_by_wordpunct):
+        for token, tag in pos_tag(tbwp_first_pass):
             if not self.is_only_punctuation(token):
                 lemmatized_token = lemmatizer.lemmatize(token, self.get_wordnet_pos(tag))
                 pos_tags.append((lemmatized_token, tag))
@@ -50,16 +40,23 @@ class Tokenizer:
                 pos_tags.append((token, "Punct"))
         return pos_tags
 
-    def _filter_and_group_tokens(self, pos_tags):
-        """Filtrage des tokens et regroupement en syntagmes."""
-        stop_words = set(stopwords.words('english'))
+    def _filter_and_group_tokens(self, pos_tags) -> tuple[ndarray, ndarray]:
+        """
+        3 responsibilities:
+        1) Filtering : remove stopwords, punctuation, empty strings
+        2) Building the tokens by word with the filtered tokens (those that have been lemmatized and tagged with POS tags in the lemmatize_and_tag_tokens method)
+        3) Grouping : group tokens by syntagm (syntagm = sequence of words that are not stopwords, punctuation or empty strings)
+
+        These 3 responsibilities were grouped to be done in one only loop, at the cost of readability.
+        """
+        stop_words = set(stopwords.words('english')) #might consider adding more languages in the future
         filtered_tokens_by_wordpunct = []
         all_syntagms = []
         current_syntagm = []
 
-        for token, pos in pos_tags:
+        for token, pos in pos_tags: #token = either a word or a punctuation sign
             if token in stop_words or self.is_only_punctuation(token) or token == '': #if this token is a "breaker"
-                if current_syntagm:
+                if current_syntagm != []:
                     all_syntagms.append(current_syntagm)
                     current_syntagm = []
                 if self.is_only_punctuation(token):
@@ -74,15 +71,13 @@ class Tokenizer:
         if current_syntagm:
             all_syntagms.append(current_syntagm)
 
-        self._tokens_by_syntagm = array(all_syntagms, dtype=object)
-        self._tokens_by_wordpunct = array(filtered_tokens_by_wordpunct)
+        return array(all_syntagms, dtype=object), array(filtered_tokens_by_wordpunct) #TODO : Think about the usage of dtype
 
-    def is_only_punctuation(self, token):
-        """Vérifie si un token est composé uniquement de ponctuation."""
+
+    def is_only_punctuation(self, token) -> bool:
         return all(char in punctuation for char in token)
 
     def get_wordnet_pos(self, tag):
-        """Convertit un tag POS en format reconnu par le lemmatizer WordNet."""
         if tag.startswith('J'):
             return wordnet.ADJ
         elif tag.startswith('V'):
@@ -90,48 +85,62 @@ class Tokenizer:
         elif tag.startswith('R'):
             return wordnet.ADV
         else:
-            return wordnet.NOUN  # Par défaut
+            return wordnet.NOUN  # Default
 
     @property
     def tokens_by_sentence(self):
         if not self._tokens_by_sentence.size:
-            print("Warning: empty list of tokens by sentence. \nPlease make sure that there's data to work with.")
+            raise TokenListIsEmpty("tokens by sentence")
         return self._tokens_by_sentence
 
     @property
     def tokens_by_wordpunct(self):
         if not self._tokens_by_wordpunct.size:
-            print("Warning: empty list of tokens by word/punctuation. \nPlease make sure that there's data to work with.")
+            raise TokenListIsEmpty("tokens by word/punctuation")
         return self._tokens_by_wordpunct
 
     @property
     def tokens_by_word(self):
         if not self._tokens_by_word.size:
-            print("Warning: empty list of tokens by word. \nPlease make sure that there's data to work with.")
+            raise TokenListIsEmpty("tokens by word")
         return self._tokens_by_word
 
     @property
     def tokens_by_syntagm(self):
         if not self._tokens_by_syntagm.size:
-            print("Warning: empty list of tokens by syntagm. \nPlease make sure that there's data to work with.")
+            raise UnprocessableTextContent("tokens by syntagm")
         return self._tokens_by_syntagm
 
     @property
     def part_of_speeches(self):
         if not self._part_of_speeches.size:
-            print("Warning: empty list of POS tags.")
+            raise TokenListIsEmpty("words' POS tags") #since POS tags are based on token by words
         return self._part_of_speeches
 
     @property
     def raw_data(self):
         if not self._raw_data:
+            raise UnprocessableTextContent("raw data")
             print("Warning: raw data is empty")
         return self._raw_data
 
+class TokenListIsEmpty(Exception):
+    """Exception raised when the token list is empty."""
+    def __init__(self, tokens_by="???"):
+        self.message = "Warning: empty list of" + tokens_by + ". \nPlease make sure that there's data to work with."
+        super().__init__(self.message)
+
+class UnprocessableTextContent(Exception):
+    def __init__(self, attribute="???") -> None:
+        self.message = "Warning: the text is senseless because :"
+        if attribute == "tokens by syntagm":
+            self.message += "Tokens by syntagm list is empty, showing that the text is either not in English or only made of stop words/punctuation."
+        elif attribute == "raw data":
+            self.message += "Raw data (e.g the text extracted from a certain source file, wether it is the user's or something scraped from the web) is empty."
+        super().__init__(self.message)
 
 
-class TokensStatsAndRearrangements: # To be referred as TSAR later on
-    #TODO : make proper getters when needed.
+class TokensStatsAndRearrangements: # To be referred as TSAR
     """
     MUST be used with ONE set of Tokens.
     - Provides statistical insights over ONE source (word frequency, text richness, POS...)
@@ -142,10 +151,10 @@ class TokensStatsAndRearrangements: # To be referred as TSAR later on
     the input and the source would allow us to conjecture (with a certain uncertainty still) if they go over the same subjects.
     This idea would help us distinguish STRUCTURAL similarities in the text from WORD similarities (since one doesn't imply the other).
     """
-    
+
     def __init__(self, base):
         self.base : Tokenizer = base
-        self.keywords_scores = {}
+        self.syntagms_scores = {} #syntagm : score
         self.bigrams = Counter(ngrams(self.base.tokens_by_word, 2))
         self.trigrams = Counter(ngrams(self.base.tokens_by_word, 3))
         self.word_freq = FreqDist(self.base.tokens_by_word)
@@ -154,29 +163,36 @@ class TokensStatsAndRearrangements: # To be referred as TSAR later on
         self.sent_lengths = {sent : len(sent.split()) for sent in self.base.tokens_by_sentence}
         self.average_sent_length = float(mean(list(self.sent_lengths.values())))
         self.median_sent_length = float(median(list(self.sent_lengths.values())))
-        self.get_syntagms_scores()
+        self.evaluate_syntagms_scores()
 
-    def get_syntagms_scores(self):
+    def evaluate_syntagms_scores(self):
+        """
+        Helper to evaluate the syntagms scores based on the RAKE algorithm.
+        Documentation : Resources/Learning Material/ResearchResources.md -> Ctrl+F "RAKE"
+        """
         word_degrees = {}
-        for s in self.base.tokens_by_syntagm: # For each syntagm
-            degree = len(s) - 1 # Length of syntagm
+        for s in self.base.tokens_by_syntagm:
+            degree = len(s) - 1
             for word in s:
                 if word not in word_degrees:
-                    word_degrees[word] = 0
+                    word_degrees[word] = degree
                 else:
                     word_degrees[word] += degree
         words_scores = {word: word_degrees[word] / self.word_freq[word] for word in self.word_freq}
         for s in self.base.tokens_by_syntagm:
             phrase_score = sum(words_scores[word] for word in s)
-            self.keywords_scores[' '.join(s)] = phrase_score
+            self.syntagms_scores[' '.join(s)] = phrase_score
+
 
 @dataclass
-class TextProcessingAlgorithms: # To be referred as TPA later on
+class TokensComparisonAlgorithms: # To be referred as TCA
     """
-    #TODO : rename that class to be more descriptive
     MUST be used with TWO sets of Tokens.
+
     Applies fundamental algorithms to evaluate similarity between TWO sources
     (like cosine similarity, n-grams similarity...) and stores the results.
+
+    Works as a "on-demand" class (meaning that the results are computed only when needed) for modularity purposes.
     """
     input_tokens_sets : TokensStatsAndRearrangements
     source_tokens_sets : TokensStatsAndRearrangements
@@ -184,11 +200,11 @@ class TextProcessingAlgorithms: # To be referred as TPA later on
     _jaccard_sim_words : float = field(init=False, default=-1)
     _cosine_sim_pos : float = field(init=False, default=-1)
     _jaccard_sim_pos : float = field(init=False, default=-1)
-    _all_term_frequencies : dict[str, float] = field(init=False, default_factory=dict) #TODO : add usage for this in phase3 to avoid repetitive computations
+    _all_term_frequencies : dict[str, float] = field(init=False, default_factory=dict)
 
-    #bigrams structure : {(word1, word2) : (input_text_occurences, source_text_occurences)}
+    #bigrams structure : {(word1, word2) : (amount_of_input_text_occurences, amount_of_source_text_occurences)}
     _identical_bigrams : dict[tuple[str, str], tuple[int, int]] = field(init=False, repr=False, default_factory=dict)
-    ##trigrams structure : {(word1, word2, word3) : (input_text_occurences, source_text_occurences)}
+    ##trigrams structure : {(word1, word2, word3) : (amount_of_input_text_occurences, amount_of_source_text_occurences)}
     _identical_trigrams : dict[tuple[str, str, str], tuple[int, int]] = field(init=False, repr=False, default_factory=dict)
 
     def __post_init__(self):
@@ -201,7 +217,7 @@ class TextProcessingAlgorithms: # To be referred as TPA later on
             self._all_term_frequencies = {term : self.input_tokens_sets.word_freq.get(term, 0) + self.source_tokens_sets.word_freq.get(term, 0) for term in all_terms}
         return self._all_term_frequencies
 
-    def cosine_similarity(self, source_terms : dict, input_terms : dict) -> float: #both args are WordFreqs (from pos or words)
+    def cosine_similarity_computation(self, source_terms : FreqDist, input_terms : FreqDist) -> float: #both args are WordFreqs (from pos or words)
         both_texts_words_union = set(source_terms.keys()).union(set(input_terms.keys()))
         input_norm = sqrt(sum(x**2 for x in input_terms.values()))
         source_norm = sqrt(sum(y**2 for y in source_terms.values()))
@@ -211,32 +227,32 @@ class TextProcessingAlgorithms: # To be referred as TPA later on
                 dot_product += source_terms[word] * input_terms[word]
         return dot_product / (input_norm * source_norm)
 
-    def jaccard_similarity(self, source_terms : set, input_terms : set) -> float:
+    def jaccard_similarity_computation(self, source_terms : set, input_terms : set) -> float:
         intersection_size = len(input_terms.intersection(source_terms))
         union_size = len(input_terms.union(source_terms))
         return intersection_size / union_size
 
     @property
     def cosine_sim_words(self):
-        return self.cosine_similarity(self.source_tokens_sets.word_freq, self.input_tokens_sets.word_freq) if self._cosine_sim_words == -1 else self._cosine_sim_words
+        return self.cosine_similarity_computation(self.source_tokens_sets.word_freq, self.input_tokens_sets.word_freq) if self._cosine_sim_words == -1 \
+        else self._cosine_sim_words
 
     @property
     def jaccard_sim_words(self):
-        return self.jaccard_similarity(set(self.source_tokens_sets.word_freq.keys()), set(self.input_tokens_sets.word_freq.keys())) if self._jaccard_sim_words == -1 else self._jaccard_sim_words
+        return self.jaccard_similarity_computation(set(self.source_tokens_sets.word_freq.keys()), set(self.input_tokens_sets.word_freq.keys())) if self._jaccard_sim_words == -1 \
+        else self._jaccard_sim_words
 
     @property
     def cosine_sim_pos(self):
-        return self.cosine_similarity(self.source_tokens_sets.pos_freq, self.input_tokens_sets.pos_freq) if self._cosine_sim_pos == -1 else self._cosine_sim_pos
+        return self.cosine_similarity_computation(self.source_tokens_sets.pos_freq, self.input_tokens_sets.pos_freq) if self._cosine_sim_pos == -1 \
+        else self._cosine_sim_pos
 
     @property
     def jaccard_sim_pos(self):
-        return self.jaccard_similarity(set(self.source_tokens_sets.pos_freq.keys()), set(self.input_tokens_sets.pos_freq.keys())) if self._jaccard_sim_pos == -1 else self._jaccard_sim_pos
+        return self.jaccard_similarity_computation(set(self.source_tokens_sets.pos_freq.keys()), set(self.input_tokens_sets.pos_freq.keys())) if self._jaccard_sim_pos == -1 \
+        else self._jaccard_sim_pos
 
     def jaccard_sim(self):
-        """
-        Same remark than for cosine sim - will maybe get moved to its own
-        dedicated function.
-        """
         if self._jaccard_sim_words == -1:
             input_words = set(self.input_tokens_sets.word_freq.keys())
             source_words = set(self.source_tokens_sets.word_freq.keys())
@@ -288,13 +304,13 @@ class TextProcessingAlgorithms: # To be referred as TPA later on
     """
 
 def extract_raw_from_file(file_path: str) -> str:
-            if not path.exists(file_path):
-                raise FileNotFoundError("Specified file doesn't exist")
+    if not path.exists(file_path):
+        raise FileNotFoundError("Specified file doesn't exist")
 
-            with open(file_path, 'rb') as file:  # Detects the encoding of file
-                raw_data = file.read()
-                result = chardet.detect(raw_data)
-                file_encoding = result['encoding']
+    with open(file_path, 'rb') as file:  # Detects the encoding of file
+        raw_data = file.read()
+        result = chardet.detect(raw_data)
+        file_encoding = result['encoding']
 
-            with open(file_path, 'r', encoding=file_encoding, errors='ignore') as f:
-                return f.read().lower()
+    with open(file_path, 'r', encoding=file_encoding, errors='ignore') as f:
+        return f.read().lower()
