+++ src/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ("TextAnalysis_AkaPhase1",)
 
-from .TextAnalysis_AkaPhase1 import Tokenizer, TextProcessingAlgorithms, TokensStatsAndRearrangements
+from .TextAnalysis_AkaPhase1 import Tokenizer, TokensComparisonAlgorithms, TokensStatsAndRearrangements
